Revert "Implement dailyArticleCount functionality"
